Move admin menu tweaks to new admin include file
