+++ public_html/wp-content/plugins/pattern-directory/bootstrap.php
@@ -8,6 +8,7 @@
 
 namespace WordPressdotorg\Pattern_Directory;
 
+require_once __DIR__ . '/includes/admin.php';
 require_once __DIR__ . '/includes/class-rest-flags-controller.php';
 require_once __DIR__ . '/includes/pattern-post-type.php';
 require_once __DIR__ . '/includes/pattern-flag-post-type.php';
